codebase cleanup
