+++ template/dcmaui_app_rewrite/ios/Runner/AppDelegate.swift
@@ -1,0 +1,17 @@
+import UIKit
+import Flutter
+import yoga
+
+@UIApplicationMain
+class AppDelegate: FlutterAppDelegate {
+    lazy var flutterEngine = FlutterEngine(name: "main engine")
+    
+   
+    override func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
+    ) -> Bool {
+        DivergeToDCMAUI()
+        return super.application(application, didFinishLaunchingWithOptions: launchOptions)
+    }
+}
