+++ template/dcmaui_app_rewrite/ios/Runner/DCMAUI/Utilities/framework/FontLoader.swift
